requirements.txt
flask==3.0.0
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
chromadb==0.4.15
pypdf==3.17.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
torch==2.1.0
numpy